+++ benchmarks/cooperative-all.ts
@@ -4,7 +4,8 @@ import { arch, platform, release } from 'node:os'
 import { performance } from 'node:perf_hooks'
 
 import { assertPromise } from '../src/Async.js'
-import { all, firstSuccess, fork, forkIn, race, withBoundedConcurrency, withCoopConcurrency, withUnboundedConcurrency } from '../src/Concurrent.js'
+import { all, firstSuccess, fork, forkIn, race, withBoundedConcurrency, withUnboundedConcurrency } from '../src/Concurrent.js'
+import { withCoopConcurrency } from '../src/experimental/concurrent/cooperative.js'
 import { Effect } from '../src/Effect.js'
 import { fail, returnFail } from '../src/Fail.js'
 import { andFinally } from '../src/Finalization.js'
