+++ src/experimental/concurrent/cooperative.ts
@@ -1,6 +1,6 @@
 import { Async } from '../../Async.js'
 import { at } from '../../Breadcrumb.js'
-import { Fork } from './effects.js'
+import { Fork } from '../../internal/concurrent/effects.js'
 import { Fail } from '../../Fail.js'
 import { flatMap, flatten, Fx, fx, ok, runPromise } from '../../Fx.js'
 import { Handle } from '../../Handler.js'
@@ -9,11 +9,11 @@ import type { CapturedHandler } from '../../HandlerCapture.js'
 import { Task } from '../../Task.js'
 import type { TraceOrigin } from '../../Trace.js'
 import { captureTrace, getTrace } from '../../Trace.js'
-import { ForkError, capturePrependTraceWithContext, originOfUnhandledFail, runtimeContextOfEffect, traceUnhandledFail, traceWithCause } from '../forkDiagnostics.js'
-import { InterruptMaskBegin, InterruptMaskEnd, InterruptMaskState } from '../interrupt.js'
-import { withInterpretedReturn } from '../iteratorClose.js'
-import { attachRuntimeContext, currentRuntimeContext, getRuntimeContext, runtimeScopeExit, RuntimeScopeExit, withActiveRuntimeContext } from '../runtimeContext.js'
-import { shouldReleaseSlotForAsync } from './cooperativeAsync.js'
+import { ForkError, capturePrependTraceWithContext, originOfUnhandledFail, runtimeContextOfEffect, traceUnhandledFail, traceWithCause } from '../../internal/forkDiagnostics.js'
+import { InterruptMaskBegin, InterruptMaskEnd, InterruptMaskState } from '../../internal/interrupt.js'
+import { withInterpretedReturn } from '../../internal/iteratorClose.js'
+import { attachRuntimeContext, currentRuntimeContext, getRuntimeContext, runtimeScopeExit, RuntimeScopeExit, withActiveRuntimeContext } from '../../internal/runtimeContext.js'
+import { shouldReleaseSlotForAsync } from '../../internal/concurrent/cooperativeAsync.js'
 
 export interface CooperativeConfig {
   readonly concurrency: number
