+++ src/exports/experimental/concurrent.ts
@@ -1,0 +1,4 @@
+export {
+  withCoopConcurrency,
+  type CoopConcurrencyOptions
+} from '../../experimental/concurrent/cooperative.js'
